+++ HISTORY.rst
@@ -85,6 +85,7 @@
 * Update group members view permission. The group edit and 
   user edit permission are now required.
 * Add keyword arguments to messages uses.
+* Add keyword arguments to the reverse use in views.
 
 3.1.11 (2019-04-XX)
 ===================